+++ spec/cli/nextify_spec.rb
@@ -10,7 +10,7 @@ describe "ruby-next nextify" do
     File.delete(File.join(__dir__, "dummy", ".rbnextrc")) if File.exist?(File.join(__dir__, "dummy", ".rbnextrc"))
   end
 
-  it "generates .rbnxt/{2.6, 2.7, 3.0, next} folders with the transpiled files required for each version" do
+  it "generates .rbnext/{2.6, 2.7, 3.0, next} folders with the transpiled files required for each version" do
     run_ruby_next "nextify #{File.join(__dir__, "dummy")} --proposed --edge" do |_status, _output, err|
       File.exist?(File.join(__dir__, "dummy", ".rbnext", "2.6", "transpile_me.rb")).should equal false
       File.exist?(File.join(__dir__, "dummy", ".rbnext", "2.6", "namespaced", "pattern_matching.rb")).should equal false
@@ -26,7 +26,7 @@ describe "ruby-next nextify" do
     end
   end
 
-  it "generates .rbnxt/{2.7, 3.0, next} folders when --min-version is provided" do
+  it "generates .rbnext/{2.7, 3.0, next} folders when --min-version is provided" do
     run_ruby_next "nextify #{File.join(__dir__, "dummy")} --proposed --edge --min-version=2.6" do |_status, _output, err|
       File.exist?(File.join(__dir__, "dummy", ".rbnext", "2.6", "transpile_me.rb")).should equal false
       File.exist?(File.join(__dir__, "dummy", ".rbnext", "2.6", "namespaced", "pattern_matching.rb")).should equal false
@@ -39,6 +39,12 @@ describe "ruby-next nextify" do
       File.exist?(File.join(__dir__, "dummy", ".rbnext", "3.0", "namespaced", "pattern_matching.rb")).should equal true
       File.exist?(File.join(__dir__, "dummy", ".rbnext", "3.0", "namespaced", "endless_nameless.rb")).should equal false
       File.exist?(File.join(__dir__, "dummy", ".rbnext", "1995.next", "method_reference.rb")).should equal true
+    end
+  end
+
+  it "generates empty .rbnext folders when nothing to transpile with .keep" do
+    run_ruby_next "nextify #{File.join(__dir__, "dummy", "namespaced")} --min-version=3.0" do |_status, _output, err|
+      File.exist?(File.join(__dir__, "dummy", "namespaced", ".rbnext", ".keep")).should equal true
     end
   end
 
@@ -66,7 +72,7 @@ describe "ruby-next nextify" do
     end
   end
 
-  it "generates .rbnxt/custom folder with versions" do
+  it "generates .rbnext/custom folder with versions" do
     run_ruby_next(
       "nextify #{File.join(__dir__, "dummy")} " \
       "--output=#{File.join(__dir__, "dummy", ".rbnext", "custom")}"
@@ -137,7 +143,6 @@ describe "ruby-next nextify" do
     end
   end
 
-  # TODO: add edge features
   it "returns error if --rewrite is provided with wrong rewriter" do
     run_ruby_next(
       "nextify #{File.join(__dir__, "dummy")} " \
@@ -149,7 +154,6 @@ describe "ruby-next nextify" do
     end
   end
 
-  # TODO: add edge features
   it "generates one version if --rewrite is provided with rewriter from edge along with --edge option" do
     run_ruby_next(
       "nextify #{File.join(__dir__, "dummy")} " \